+++ data/sample/initialize_files/sample_local_environment.ini
@@ -1,7 +1,7 @@
 [MAGNETIC_FIELD_ENVIRONMENT]
 calculation = ENABLE
 logging = ENABLE
-coefficient_file = ../../../s2e-core/src/library/igrf/igrf13.coef
+coefficient_file = ../../../s2e-core/src/library/external/igrf/igrf13.coef
 magnetic_field_random_walk_speed_nT = 10.0
 magnetic_field_random_walk_limit_nT = 400.0
 magnetic_field_white_noise_standard_deviation_nT = 50.0
