+++ src/environment/global/gnss_satellites.cpp
@@ -5,8 +5,8 @@
 
 #include "gnss_satellites.hpp"
 
-#include <library/sgp4/sgp4ext.h>   //for jday()
-#include <library/sgp4/sgp4unit.h>  //for gstime()
+#include <library/external/sgp4/sgp4ext.h>   //for jday()
+#include <library/external/sgp4/sgp4unit.h>  //for gstime()
 
 #include <algorithm>
 #include <environment/global/physical_constants.hpp>
