+++ src/library/external/igrf/igrf11.coef
