+++ src/library/external/igrf/igrf12.coef
