+++ src/library/external/igrf/igrf13.coef
